Separate out fw and project eslint config
