+++ packages/cli/src/commands/test.js
@@ -103,7 +103,7 @@ export const handler = async ({
     collectCoverage ? '--collectCoverage' : null,
     '--passWithNoTests',
     ...jestFilterArgs,
-    sides.includes('api') && '--runInBand',
+    '--runInBand', // @TODO always run in band, even for web as we get babel errors https://github.com/redwoodjs/redwood/issues/3646
   ].filter((flagOrValue) => flagOrValue !== null) // Filter out nulls, not booleans because user may have passed a --something false flag
 
   // If the user wants to watch, set the proper watch flag based on what kind of repo this is
